Dynamics jam (3)
Updated comments, revamped the dynamics layout with the inclusion of rotation matrices and an intuitive method of labelling vectors expressed in different frames (all sharing a common origin). Finished filling out the dynamics model in its general form (writing in the equations with general expressions for lift, thrust, drag, and gravitational acceleration as well as Euler's law with the body-fixed frame torques). Successfully finished evaluating the gradient of the gravitational potential, but found a bug in step 2 of correcting the geopotential coefficients (unresolved). Changed the reference frame for orientation from the noninertial rail to the inertial ECI frame, which isn't very useful in itself, but results are easily transformed to other frames afterwards. Introduced several functions providing a standard library of rotation matrices.