Add cache system for loading Earth model
Cache system saves/reads a small .mat file containing parameters
defining the Earth model that is compared against current defining
parameters. Those parameters that are the same simply reuse the model
defined from the previous run of rtraj.m.
Note: The working directory currently must be that of rtraj.m for the
cache to work properly.