Add body-velocity transformation & angle of attack
The transformation between the body-fixed frame and the velocity frame
is obtained by expressing the (noninertial) body's translational ENV
velocity in the body-fixed frame and calculating the associated angle of
attack and 'velocity roll' angles. These angles provide a spherical
coordinate description of the body-fixed velocity (the velocity parallel
to the freestream - no wind yet). The transformation then follows from
2 elementary rotations (3-2 for body-to-velocity frame).